QChar: move definitions of inline methods into the class body
This hopefully prevents buggy compilers from choking on the definitions
outside the class as inconsistent, even though they are consistent.

Also helps developers, as the rather lengthy attribute lists for
C++11 functions does not need to be repeated.

This is in preparation of plastering the QChar API with Q_DECL_NOTHROW.

I had to keep operator== and operator< outside the class because
Clang on OS X 10 didn't find it when called with types implicitly
convertible to QChar.

Change-Id: I8bd634bc756d1bf5d208d7a2ce7e7ded73c4ce56
Reviewed-by: Olivier Goffart (Woboq GmbH) <ogoffart@woboq.com>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>