QFileSystemModel: reimplement sibling for same-row
Since QFSM, like most models, uses QModelIndex::internalPointer() as an
indicator of the row, getting a sibling in the same row is as easy as
copying the internalPointer() of the incoming index and just creating
a new index with the column adjusted.

For rows, the situation is quite a bit more complicated, so we
currently continue to call the generic implementation.

Change-Id: I36921e3f9c01c458a75aa439018f21c4c657e1cf
Reviewed-by: David Faure <david.faure@kdab.com>